added a call to "lhefinitemasses" in Born.f.
This doesn't break backward compatibility for HJ, as the
quark masses are all set to zero by default.

For MiNNLOPS, we want to really minimize the number of events that the
shower can't deal with, hence we want to put b and c momenta on mass shell.
The definition of the masses will be included in files inside HJMiNNLO, so
backward compatibility is guaranteed



git-svn-id: svn://powhegbox.mib.infn.it/trunk/User-Processes-V2@3742 9e129b38-c2c0-4eec-9301-dc54fda404de