HJ/HJMiNNLO/init_couplings.f Wj/WjMiNNLO/init_couplings.f Zj/ZjMiNNLO/init_couplings.f
Added masses of b and c quarks to reshuffle kinematics and put final state b and c quarks on mass shell before the event is written on the LH file.
It only affects this last step of event generation.
Needed to avoid PY errors (PY needs hard events with b and c-quarks with nonzero mass)




git-svn-id: svn://powhegbox.mib.infn.it/trunk/User-Processes-V2@3743 9e129b38-c2c0-4eec-9301-dc54fda404de